split backup test into two binaries/containers
