+++ installation/scripts/e2e-testing.sh
@@ -32,6 +32,9 @@ cleanupHelmE2ERelease () {
 # creates a config map which provides the testing bundles	
 if [[ "${ACTION}" == "testBeforeBackup" ]]; then
   injectTestingAddons
+  job=before
+else
+  job=after
 fi  
 
 testcase="${ROOT_PATH}"/../../tests/end-to-end/backup/chart/backup-test
@@ -40,7 +43,7 @@ release=$(basename "$testcase")
 ADMIN_EMAIL=$(kubectl get secret admin-user -n kyma-system -o jsonpath="{.data.email}" | base64 --decode)
 ADMIN_PASSWORD=$(kubectl get secret admin-user -n kyma-system -o jsonpath="{.data.password}" | base64 --decode)
 
-helm install "$testcase" --name "${release}" --namespace backup-test --set global.ingress.domainName="${DOMAIN}" --set action="${ACTION}" --set-file global.adminEmail=<(echo -n "${ADMIN_EMAIL}") --set-file global.adminPassword=<(echo -n "${ADMIN_PASSWORD}") --tls
+helm install "$testcase" --name "${release}" --namespace backup-test --set global.ingress.domainName="${DOMAIN}" --set job="${job}" --set-file global.adminEmail=<(echo -n "${ADMIN_EMAIL}") --set-file global.adminPassword=<(echo -n "${ADMIN_PASSWORD}") --tls
 
 suiteName="testsuite-backup-$(date '+%Y-%m-%d-%H-%M')"
 echo "---------------------------------------------------"
