+++ tests/end-to-end/backup/pkg/config/config.go
@@ -3,6 +3,7 @@ package config
 import (
 	"os"
 
+	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 )
